# contracts/openapi/v1.yaml
openapi: 3.1.0
info:
  title: BharatStudio Alerts API
  version: 1.0.0-draft
  description: >-
    Draft v1 REST contract for BharatStudio Alerts and Companion. This contract
    is not approved for production until L01 review and verification pass.
servers:
  - url: https://api.example.invalid
    description: Placeholder only; production domain is not yet approved.
...
paths:
...
  /v1/channels/{channelId}/overlay/session:
    post:
      tags: [Overlay]
      operationId: createOverlaySession
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/ChannelId'
      responses:
        '201':
          description: Short-lived scoped overlay session.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OverlaySession'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /v1/overlays/{overlayId}:
    delete:
      tags: [Overlay]
      operationId: revokeOverlaySession
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/OverlayId'
      responses:
        '204':
          description: Overlay session revoked.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/overlays/{overlayId}/rotate:
    post:
      tags: [Overlay]
      operationId: rotateOverlaySession
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/OverlayId'
      responses:
        '201':
          description: Replacement scoped overlay session; previous session is revoked.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OverlaySession'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/overlays/{overlayId}/events:
    get:
      tags: [Overlay]
      operationId: streamOverlayEvents
      security: [{ overlayToken: [] }]
      parameters:
        - name: overlayId
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: Last-Event-ID
          in: header
          required: false
          description: Durable replay cursor returned by the previous SSE event.
          schema: { type: string, maxLength: 256 }
        - name: limit
          in: query
          required: false
          schema: { type: integer, minimum: 1, maximum: 100, default: 50 }
      responses:
        '200':
          description: Server-sent overlay events.
          headers:
            Cache-Control:
              schema: { type: string, const: no-store }
            X-Overlay-Stream-Version:
              schema: { type: string }
          content:
            text/event-stream:
              schema:
                type: string
                description: SSE frames whose data validates as OverlaySseEvent.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '410':
          description: Cursor is outside the replay window; client must resync.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /v1/overlays/{overlayId}/cursor:
    post:
      tags: [Overlay]
      operationId: acknowledgeOverlayCursor
      security: [{ overlayToken: [] }]
      parameters:
        - $ref: '#/components/parameters/OverlayId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OverlayCursorAcknowledgement'
      responses:
        '204':
          description: Cursor acknowledgement stored.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '503':
          $ref: '#/components/responses/RetryableFailure'
...
  /v1/overlay-lottie/{overlayId}:
    get:
      tags: [Overlay]
      operationId: listOverlayLottieAssets
      security: [{ overlayToken: [] }]
      parameters: [{ $ref: '#/components/parameters/OverlayId' }]
      responses:
        '200':
          description: Token-scoped Lottie asset references.
          content: { application/json: { schema: { $ref: '#/components/schemas/OverlayLottieListResponse' } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '503': { $ref: '#/components/responses/RetryableFailure' }
  /v1/overlay-lottie/{overlayId}/{artifactId}:
    get:
      tags: [Overlay]
      operationId: getOverlayLottieAsset
      security: [{ overlayToken: [] }]
      parameters:
        - $ref: '#/components/parameters/OverlayId'
        - $ref: '#/components/parameters/ArtifactId'
      responses:
        '200':
          description: Token-scoped Lottie bytes; private and never cacheable.
          headers: { Cache-Control: { schema: { type: string, const: 'private, no-store' } } }
          content: { application/json: { schema: { type: object } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '503': { $ref: '#/components/responses/RetryableFailure' }
  /v1/overlay-audio/{overlayId}/{artifactId}:
    get:
      tags: [Overlay]
      operationId: getOverlayAudioAsset
      security: [{ overlayToken: [] }]
      parameters:
        - $ref: '#/components/parameters/OverlayId'
        - $ref: '#/components/parameters/ArtifactId'
      responses:
        '200':
          description: Token-scoped audio bytes; private and never cacheable.
          headers: { Cache-Control: { schema: { type: string, const: 'private, no-store' } } }
          content: { audio/wav: { schema: { type: string, format: binary } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '503': { $ref: '#/components/responses/RetryableFailure' }
...
components:
  parameters:
...
    ChannelId:
      name: channelId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    OverlayId:
      name: overlayId
      in: path
      required: true
      schema: { type: string, format: uuid }
    ArtifactId:
      name: artifactId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque-session-token
    viewerBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque-viewer-session-token
    overlayToken:
      type: http
      scheme: bearer
      bearerFormat: opaque-overlay-token
    serviceIdentity:
      type: http
      scheme: bearer
      bearerFormat: OIDC-service-identity-token
  responses:
    BadRequest:
      description: Request validation failed.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    Forbidden:
      description: Authenticated identity lacks permission.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...
    RetryableFailure:
      description: Temporary service failure; the client may retry.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...
    Unauthorized:
      description: Missing or invalid authentication.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
  schemas:
...
    OverlayLottieAssetRef:
      type: object
      additionalProperties: false
      required: [displayStyle, artifactId]
      properties:
        displayStyle: { type: string, enum: [small_pill, compact_card, standard_card, large_card, banner, celebration] }
        artifactId: { type: string, format: uuid }
    OverlayLottieListResponse:
      type: object
      additionalProperties: false
      required: [schemaVersion, items]
      properties:
        schemaVersion: { type: string, const: v1 }
        items: { type: array, maxItems: 6, items: { $ref: '#/components/schemas/OverlayLottieAssetRef' } }
...
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [schemaVersion, errorCode, message, traceId]
      properties:
        schemaVersion: { type: string, const: v1 }
        errorCode: { type: string, minLength: 1, maxLength: 80 }
        message: { type: string, minLength: 1, maxLength: 240 }
        traceId: { type: string, minLength: 1, maxLength: 128 }
        retryable: { type: boolean }
...
    OverlaySession:
      type: object
      additionalProperties: false
      required: [schemaVersion, overlayId, expiresAt, streamUrl]
      properties:
        schemaVersion: { type: string, const: v1 }
        overlayId: { type: string, format: uuid }
        expiresAt: { type: string, format: date-time }
        streamUrl:
          type: string
          format: uri
          description: >-
            Web overlay HTML URL for an OBS Browser Source. The short-lived
            credential must be carried only in the URL fragment; this URL must
            contain no query credential and the API SSE endpoint must be
            reached by the page with an Authorization bearer header.
...
    OverlayCursorAcknowledgement:
      type: object
      additionalProperties: false
      required: [cursor, eventId]
      properties:
        cursor: { type: string, minLength: 1, maxLength: 256 }
        eventId: { type: string, format: uuid }
...